requirements.txt
asgiref==3.8.0
Django==5.0.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1
